refactor: change structure albums, artists module
