+++ src/index.ts
@@ -1,14 +1,13 @@
 import 'dotenv/config';
 import { ApolloServer, gql } from 'apollo-server';
 import { allSchemas } from './modules/allSchemas.js';
-import { artistsResolver } from './modules/artists/resolver.js';
 import { allServices } from './modules/allServices.js';
-import { albumResolver } from './modules/albums/resolver.js';
 import { bandsResolver } from './modules/bands/resolver.js';
 import { genresResolver } from './modules/genres/resolver.js';
 import { trackResolver } from './modules/tracks/resolver.js';
 import { usersResolver } from './modules/users/resolver.js';
 import { jwtResolver } from './modules/jwt/resolver.js';
+import { allQueries } from './modules/allQueries.js';
 
 const PORT: number = +process.env.PORT || 3000;
 
@@ -17,8 +16,7 @@ const typeDefs = gql`
 `;
 const resolvers = {
   Query: {
-    ...artistsResolver,
-    ...albumResolver,
+    ...allQueries,
     ...bandsResolver,
     ...genresResolver,
     ...trackResolver,
