+++ src/modules/albums/album.schema.ts
